kernel/thread: Cast atomic_store_explicit arg to allow including header in C++
The first two arguments to atomic_store_explicit must have
the same type which previously wasn't the case because the
SMP_MAX_CPUS macros defined in the build system isn't an
int32_t literal.

Bug: 230134581

+++ include/kernel/thread.h
@@ -245,7 +245,7 @@ static inline void thread_lock_complete(void) {
 static inline void thread_unlock_prepare(void) {
     DEBUG_ASSERT(arch_ints_disabled());
     DEBUG_ASSERT(thread_lock_owner_get() == arch_curr_cpu_num());
-    atomic_store_explicit(&thread_lock_owner, SMP_MAX_CPUS,
+    atomic_store_explicit(&thread_lock_owner, (uint)SMP_MAX_CPUS,
                           memory_order_relaxed);
 }
